requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
openai==1.104.2
pandas==2.3.2
python-docx==1.2.0
openpyxl==3.1.5
jinja2==3.1.4
